iLab testing

+++ libnetfiles.c
@@ -258,7 +258,7 @@ ssize_t netread(int fd, void * buf, size_t nbyte) {
 
 	// Read character into buffer
 	readn(sockfd, (char *)buf, nbyte);
-	printf("Client: netread: %s %zd\n", buf, nbyte);
+	printf("Client: netread: %s %zd\n", buf, (size_t)nbyte);
 
 	// Receive response from server
 	Command_packet * packet = (Command_packet *)readCommand(sockfd);
